fix(MyPage): 닉네임 컴포넌트를 분리하고 주석 정리와 안쓰는 코드를 삭제합니다. by kimnamgyu93
